Update all build chain to make it more efficient

+++ packages/clonezilla/build.sh
@@ -1,26 +1,32 @@
 CURRENT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 source $CURRENT_DIR/version.sh
 
-set -x
-if [ ! -f $working_directory/sources/$clonezilla_iso ]; then
-    wget -P $working_directory/sources/ $clonezilla_iso_url
-fi
+if [ ! -f $tags_directory/clonezilla-$distribution-$distribution_version-$clonezilla_version ]; then
 
-rm -Rf $working_directory/build/clonezilla
-mkdir -p $working_directory/build/clonezilla
-cd $working_directory/build/clonezilla
-mkdir clonezilla-$clonezilla_version
+    set -x
+    if [ ! -f $working_directory/sources/$clonezilla_iso ]; then
+        wget -P $working_directory/sources/ $clonezilla_iso_url
+    fi
 
-osirrox -indev $working_directory/sources/$clonezilla_iso -extract /live clonezilla-$clonezilla_version/live/
-$(which cp) -af $root_directory/clonezilla/* clonezilla-$clonezilla_version/
-tar cvzf clonezilla.tar.gz clonezilla-$clonezilla_version
-rpmbuild -ta clonezilla.tar.gz --target=noarch --define "_software_version $clonezilla_version"
+    rm -Rf $working_directory/build/clonezilla
+    mkdir -p $working_directory/build/clonezilla
+    cd $working_directory/build/clonezilla
+    mkdir clonezilla-$clonezilla_version
 
-if [ $distribution == "Ubuntu" ] || [ $distribution == "Debian" ]; then
-    cd /root
-    alien --to-deb --scripts /root/rpmbuild/RPMS/noarch/clonezilla-*
-    mkdir -p /root/debbuild/DEBS/noarch/
-    mv *.deb /root/debbuild/DEBS/noarch/
+    osirrox -indev $working_directory/sources/$clonezilla_iso -extract /live clonezilla-$clonezilla_version/live/
+    $(which cp) -af $root_directory/clonezilla/* clonezilla-$clonezilla_version/
+    tar cvzf clonezilla.tar.gz clonezilla-$clonezilla_version
+    rpmbuild -ta clonezilla.tar.gz --target=noarch --define "_software_version $clonezilla_version"
+
+    if [ $distribution == "Ubuntu" ] || [ $distribution == "Debian" ]; then
+        cd /root
+        alien --to-deb --scripts /root/rpmbuild/RPMS/noarch/clonezilla-*
+        mkdir -p /root/debbuild/DEBS/noarch/
+        mv *.deb /root/debbuild/DEBS/noarch/
+    fi
+    # Build success, tag it
+    touch $tags_directory/clonezilla-$distribution-$distribution_version-$clonezilla_version
+
 fi
 
 set +x
